// src/interfaces/interface.tsx
 export interface ISignupState {
    username: string;
    email: string;
    password: string;
    confirmPassword: string;
    Specialization: string;
    clinic_address: string;
    national_id: string;
    id_card :string ;
    Personal_photograph: string; 
    phoneNumber: string;
    age : string;   
    Gender : string;
}